src/backend/requirements/requirements.txt
django==3.1.7
psycopg2-binary==2.8.6
neo4j==4.2.1
py2neo==2021.0.1
django-import-export==2.5.0
djangorestframework==3.12.2
djangorestframework-filters
djangorestframework-simplejwt==3.2.3
djangorestframework-api-key==2.*
boto3
